fixing reg template

+++ cadettracker/views.py
@@ -31,11 +31,12 @@ def company(request,company_id):
                                         request))
 
 def fulfillRequest(request, request_id):
-    fulfill = CompanyNeedsSupply.objects.get(pk=request_id)
-    fulfill.delete()
-    # template = loader.get_template('cadettracker/regiment.html')
-    return HttpResponseRedirect('/supply')
-    return render(request, 'cadettracker/delete.html')
+    if request.method == "POST":
+        fulfill = CompanyNeedsSupply.objects.get(pk=request_id)
+        fulfill.delete()
+        # template = loader.get_template('cadettracker/regiment.html')
+        return HttpResponseRedirect('/supply')
+        # return render(request, 'cadettracker/delete.html')
 
 def reg(request, reg_id):
     regi = Regiment.objects.get(pk=reg_id)
